move common setup and teardown into conftest file
